Revert "Merge pull request #64 from ashishsingh18/PerfusionAlignmentChart"
This reverts commit 8b9aa06c87805fa9604dd06bc4178ff3ec7bab19, reversing
changes made to cc50abea4ae822b0b19dd2f7ced2c49a20e58e9a.
